Fixed demo sync and renamed it to demo_ipython
